# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Personal finance management API for tracking transactions, budgets, savings goals, and debts.",
        "title": "WealthPath API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@wealthpath.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/recurring": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get all recurring transactions for the current user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "List recurring transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new recurring transaction (income or expense)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "Create a recurring transaction",
                "parameters": [
                    {
                        "description": "Recurring transaction data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_service.CreateRecurringInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recurring/upcoming": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get the next 10 upcoming recurring transactions",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "Get upcoming recurring transactions",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recurring/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get a recurring transaction by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "Get a recurring transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recurring Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing recurring transaction",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "Update a recurring transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recurring Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated recurring transaction data",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_service.UpdateRecurringInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a recurring transaction by ID",
                "tags": [
                    "recurring"
                ],
                "summary": "Delete a recurring transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recurring Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recurring/{id}/pause": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Pause a recurring transaction to stop automatic generation",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "Pause a recurring transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recurring Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/recurring/{id}/resume": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Resume a paused recurring transaction",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "recurring"
                ],
                "summary": "Resume a recurring transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Recurring Transaction ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringTransaction"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_wealthpath_backend_internal_model.RecurringFrequency": {
            "type": "string",
            "enum": [
                "daily",
                "weekly",
                "biweekly",
                "monthly",
                "yearly"
            ],
            "x-enum-varnames": [
                "FrequencyDaily",
                "FrequencyWeekly",
                "FrequencyBiweekly",
                "FrequencyMonthly",
                "FrequencyYearly"
            ]
        },
        "github_com_wealthpath_backend_internal_model.RecurringTransaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringFrequency"
                },
                "id": {
                    "type": "string"
                },
                "isActive": {
                    "type": "boolean"
                },
                "lastGenerated": {
                    "type": "string"
                },
                "nextOccurrence": {
                    "type": "string"
                },
                "startDate": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.TransactionType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_model.TransactionType": {
            "type": "string",
            "enum": [
                "income",
                "expense"
            ],
            "x-enum-varnames": [
                "TransactionTypeIncome",
                "TransactionTypeExpense"
            ]
        },
...
        "github_com_wealthpath_backend_internal_service.CreateRecurringInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringFrequency"
                },
                "startDate": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.TransactionType"
                }
            }
        },
...
        "github_com_wealthpath_backend_internal_service.UpdateRecurringInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "category": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "endDate": {
                    "type": "string"
                },
                "frequency": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.RecurringFrequency"
                },
                "isActive": {
                    "type": "boolean"
                },
                "startDate": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/github_com_wealthpath_backend_internal_model.TransactionType"
                }
            }
        },
...
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}